refactor(workstation): extract AI-draft + pull-request actions into hooks (0.72 app.ts decomposition)
